+++ Doctrine/Orm/Filter/DateFilter.php
@@ -14,6 +14,7 @@ namespace Dunglas\ApiBundle\Doctrine\Orm\Filter;
 use Doctrine\Common\Persistence\ManagerRegistry;
 use Doctrine\ORM\QueryBuilder;
 use Dunglas\ApiBundle\Api\ResourceInterface;
+use Dunglas\ApiBundle\Doctrine\Orm\Util\QueryUtils;
 use Symfony\Component\HttpFoundation\RequestStack;
 
 /**
@@ -62,8 +63,6 @@ class DateFilter extends AbstractFilter
      */
     public function apply(ResourceInterface $resource, QueryBuilder $queryBuilder)
     {
-        $fieldNames = $this->getDateFieldNames($resource);
-
         $request = $this->requestStack->getCurrentRequest();
         if (null === $request) {
             return;
@@ -71,20 +70,43 @@ class DateFilter extends AbstractFilter
 
         foreach ($this->extractProperties($request) as $property => $values) {
             // Expect $values to be an array having the period as keys and the date value as values
-            if (!isset($fieldNames[$property]) || !is_array($values) || !$this->isPropertyEnabled($property)) {
+            if (
+                !$this->isPropertyEnabled($property)
+                || !$this->isPropertyMapped($property, $resource)
+                || !$this->isDateField($property, $resource)
+                || !is_array($values)
+            ) {
                 continue;
+            }
+
+            $alias = 'o';
+            $field = $property;
+
+            if ($this->isPropertyNested($property)) {
+                $propertyParts = $this->splitPropertyParts($property);
+
+                $parentAlias = $alias;
+
+                foreach ($propertyParts['associations'] as $association) {
+                    $alias = QueryUtils::generateJoinAlias($association);
+                    $queryBuilder->join(sprintf('%s.%s', $parentAlias, $association), $alias);
+                    $parentAlias = $alias;
+                }
+
+                $field = $propertyParts['field'];
             }
 
             $nullManagement = isset($this->properties[$property]) ? $this->properties[$property] : null;
 
             if (self::EXCLUDE_NULL === $nullManagement) {
-                $queryBuilder->andWhere($queryBuilder->expr()->isNotNull(sprintf('o.%s', $property)));
+                $queryBuilder->andWhere($queryBuilder->expr()->isNotNull(sprintf('%s.%s', $alias, $field)));
             }
 
             if (isset($values[self::PARAMETER_BEFORE])) {
                 $this->addWhere(
                     $queryBuilder,
-                    $property,
+                    $alias,
+                    $field,
                     self::PARAMETER_BEFORE,
                     $values[self::PARAMETER_BEFORE],
                     $nullManagement
@@ -94,7 +116,8 @@ class DateFilter extends AbstractFilter
             if (isset($values[self::PARAMETER_AFTER])) {
                 $this->addWhere(
                     $queryBuilder,
-                    $property,
+                    $alias,
+                    $field,
                     self::PARAMETER_AFTER,
                     $values[self::PARAMETER_AFTER],
                     $nullManagement
@@ -104,44 +127,39 @@ class DateFilter extends AbstractFilter
     }
 
     /**
-     * Adds the where clause accordingly to the choosed null management.
+     * Adds the where clause according to the chosen null management.
      *
      * @param QueryBuilder $queryBuilder
-     * @param string       $property
-     * @param string       $parameter
+     * @param string       $alias
+     * @param string       $field
+     * @param string       $operator
      * @param string       $value
      * @param int|null     $nullManagement
      */
-    private function addWhere(QueryBuilder $queryBuilder, $property, $parameter, $value, $nullManagement)
+    private function addWhere(QueryBuilder $queryBuilder, $alias, $field, $operator, $value, $nullManagement)
     {
-        $queryParameter = sprintf('date_%s_%s', $parameter, $property);
-        $where = sprintf('o.%s %s= :%s', $property, self::PARAMETER_BEFORE === $parameter ? '<' : '>', $queryParameter);
-
-        $queryBuilder->setParameter($queryParameter, new \DateTime($value));
+        $valueParameter = QueryUtils::generateParameterName(sprintf('%s_%s', $field, $operator));
+        $baseWhere = sprintf('%s.%s %s :%s', $alias, $field, self::PARAMETER_BEFORE === $operator ? '<=' : '>=', $valueParameter);
 
         if (null === $nullManagement || self::EXCLUDE_NULL === $nullManagement) {
-            $queryBuilder->andWhere($where);
-
-            return;
-        }
-
-        if (
-            (self::PARAMETER_BEFORE === $parameter && self::INCLUDE_NULL_BEFORE === $nullManagement)
+            $queryBuilder->andWhere($baseWhere);
+        } elseif (
+            (self::PARAMETER_BEFORE === $operator && self::INCLUDE_NULL_BEFORE === $nullManagement)
             ||
-            (self::PARAMETER_AFTER === $parameter && self::INCLUDE_NULL_AFTER === $nullManagement)
+            (self::PARAMETER_AFTER === $operator && self::INCLUDE_NULL_AFTER === $nullManagement)
         ) {
             $queryBuilder->andWhere($queryBuilder->expr()->orX(
-                $where,
-                $queryBuilder->expr()->isNull(sprintf('o.%s', $property))
+                $baseWhere,
+                $queryBuilder->expr()->isNull(sprintf('%s.%s', $alias, $field))
             ));
-
-            return;
+        } else {
+            $queryBuilder->andWhere($queryBuilder->expr()->andX(
+                $baseWhere,
+                $queryBuilder->expr()->isNotNull(sprintf('%s.%s', $alias, $field))
+            ));
         }
 
-        $queryBuilder->andWhere($queryBuilder->expr()->andX(
-            $where,
-            $queryBuilder->expr()->isNotNull(sprintf('o.%s', $property))
-        ));
+        $queryBuilder->setParameter($valueParameter, new \DateTime($value));
     }
 
     /**
@@ -150,11 +168,19 @@ class DateFilter extends AbstractFilter
     public function getDescription(ResourceInterface $resource)
     {
         $description = [];
-        foreach ($this->getClassMetadata($resource)->getFieldNames() as $fieldName) {
-            if ($this->isPropertyEnabled($fieldName)) {
-                $description += $this->getFilterDescription($fieldName, self::PARAMETER_BEFORE);
-                $description += $this->getFilterDescription($fieldName, self::PARAMETER_AFTER);
+
+        $properties = $this->properties;
+        if (null === $properties) {
+            $properties = array_fill_keys($this->getClassMetadata($resource)->getFieldNames(), null);
+        }
+
+        foreach ($properties as $property => $nullManagement) {
+            if (!$this->isPropertyMapped($property, $resource)) {
+                continue;
             }
+
+            $description += $this->getFilterDescription($property, self::PARAMETER_BEFORE);
+            $description += $this->getFilterDescription($property, self::PARAMETER_AFTER);
         }
 
         return $description;
@@ -163,16 +189,16 @@ class DateFilter extends AbstractFilter
     /**
      * Gets filter description.
      *
-     * @param string $fieldName
+     * @param string $property
      * @param string $period
      *
      * @return array
      */
-    private function getFilterDescription($fieldName, $period)
+    private function getFilterDescription($property, $period)
     {
         return [
-            sprintf('%s[%s]', $fieldName, $period) => [
-                'property' => $fieldName,
+            sprintf('%s[%s]', $property, $period) => [
+                'property' => $property,
                 'type' => '\DateTime',
                 'required' => false,
             ],
@@ -180,23 +206,17 @@ class DateFilter extends AbstractFilter
     }
 
     /**
-     * Gets names of fields with a date type.
+     * Determines whether the given property refers to a date field.
      *
-     * @param ResourceInterface $resource
+     * @param string $property
      *
-     * @return array
+     * @return bool
      */
-    private function getDateFieldNames(ResourceInterface $resource)
+    private function isDateField($property, ResourceInterface $resource)
     {
-        $classMetadata = $this->getClassMetadata($resource);
-        $dateFieldNames = [];
+        $propertyParts = $this->splitPropertyParts($property);
+        $metadata = $this->getNestedMetadata($resource, $propertyParts['associations']);
 
-        foreach ($classMetadata->getFieldNames() as $fieldName) {
-            if (isset(self::$doctrineDateTypes[$classMetadata->getTypeOfField($fieldName)])) {
-                $dateFieldNames[$fieldName] = true;
-            }
-        }
-
-        return $dateFieldNames;
+        return isset(self::$doctrineDateTypes[$metadata->getTypeOfField($propertyParts['field'])]);
     }
 }